Initialize evaluations in backMigrate
The proceedings page is relying on these evaluations to generate the
page. The issue previously was that there was some referencing issue
causing all of the evaluations for all tools to be appended and pushed
to the new tool to be evaluated.

That combined with all of the identical generation dates I think caused
it to also be sorted by alphabetical order so it was always reading the
same evaluation as latest.